refactor(concerto): shared useResolvedTheme hook for both renderer roots
The cadenza HUD root re-derived a lossy subset of App's theme resolution
(custom + built-in only), returning undefined for plugin-contributed theme ids.
Extract the canonical resolver (custom / built-in / plugin theme / dracula
fallback) into useResolvedTheme() and use it from both App.tsx and cadenzaHud.tsx.
Removes the duplication and fixes the latent HUD plugin-theme bug. Addresses the
provider-shell half of review finding #9.
